chore(all): init brickyard3
this brickyard3 apply new implementation of the chore framework.
for now, it has accomplished following features:

- commandLoader.js -- read the arguments from command line, and
transform them into target data type, specified in `command.yaml`
- configLoader.js -- search for the config file for global usage.
- pluginsLoader.js -- load target/all plugins' configurations and
generate an object for every plugin
- programsLoader.js -- load the assemble recipe which contains
required plugins and some special config for a particular program.
- packageInfo.js -- load the package.json of current repo.
- runtime.js -- manage the modules above to generate a runtime
configuration that can drive the program to work properly.

there are still two module left:
- bowerInstaller.js - to check the installed packages of target dir
and install the missing/mismatch bower dependencies.
- npmInstaller.js - to check the installed npm packages and install
the missing/mismatch npm dependencies.

and a bunch of stuffes lyging on the root of the repo which are
commonly required in a project.

and of cause, there exist some test spec files and the corresponding
test resources. And the current features already has their test
suites.